fix: rename filenames to follow consistent pattern
