+++ src/components/checkin/checkin_rating.tsx
