+++ src/components/checkin/index.ts
@@ -1,3 +1,3 @@
-export { default as CheckInRatings } from './checkin-rating';
-export { default as WeeklyCheckinForm } from './weekly-checkin';
-export { default as HighFive } from './high-five';
+export { default as CheckInRatings } from './checkin_rating';
+export { default as WeeklyCheckinForm } from './weekly_checkin';
+export { default as HighFive } from './high_five';
