+++ src/components/index.ts
@@ -3,5 +3,5 @@ export { default as Layout } from './layout';
 export { default as sidebar } from './sidebar';
 export { default as Topbar } from './topbar';
 export { default as LoginCard } from './login';
-export { default as ProfileBar } from './profilebar-card';
+export { default as ProfileBar } from './profilebar_card';
 export * from './checkin';
